Added sync worker logic for category and item

+++ app/src/main/java/com/fx/inventory/data/models/Response.kt
@@ -1,3 +1,3 @@
 package com.fx.inventory.data.models
 
-data class Response(var message: String?,var success:Boolean = false, var id: Long? = null)
+data class Response(var message: String?,var success:Boolean = false, var id: Int? = null)
